Check if attribute is already set in user input.
Save user data if starter conversation is triggered. Get additional data from facebook if possible.
Pre-populate conversation vars with variables in command.variables array.
If checkIfAttributeExists is enabled in user input, first check if the attribute is already set and if it is set then don't ask user for input.